Refactor 'toCamelCase' code
